+++ src/node_report.h
@@ -10,7 +10,7 @@
 #endif
 #endif
 
-namespace nodereport {
+namespace node-report {
 
 using v8::Isolate;
 using v8::Local;
@@ -21,7 +21,7 @@ using v8::Number;
 using v8::String;
 using v8::Value;
 
-// Bit-flags for NodeReport trigger options
+// Bit-flags for node-report trigger options
 #define NR_EXCEPTION  0x01
 #define NR_FATALERROR 0x02
 #define NR_SIGNAL     0x04
@@ -33,14 +33,14 @@ using v8::Value;
 
 enum DumpEvent {kException, kFatalError, kSignal_JS, kSignal_UV, kJavaScript};
 
-void TriggerNodeReport(Isolate* isolate, DumpEvent event, const char* message, const char* location, char* name);
+void Triggernode-report(Isolate* isolate, DumpEvent event, const char* message, const char* location, char* name);
 
-unsigned int ProcessNodeReportEvents(const char* args);
-unsigned int ProcessNodeReportCoreSwitch(const char* args);
-unsigned int ProcessNodeReportSignal(const char* args);
-void ProcessNodeReportFileName(const char* args);
-void ProcessNodeReportDirectory(const char* args);
-unsigned int ProcessNodeReportVerboseSwitch(const char* args);
+unsigned int ProcessNode_ReportEvents(const char* args);
+unsigned int ProcessNode_ReportCoreSwitch(const char* args);
+unsigned int ProcessNode_ReportSignal(const char* args);
+void ProcessNode_ReportFileName(const char* args);
+void ProcessNode_ReportDirectory(const char* args);
+unsigned int ProcessNode_ReportVerboseSwitch(const char* args);
 
 void SetLoadTime();
 void SetVersionString(Isolate* isolate);
@@ -78,6 +78,6 @@ inline static int snprintf(char* buffer, size_t n, const char* format, ...) {
 #define __func__ __FUNCTION__
 #endif  // defined( _MSC_VER ) && (_MSC_VER < 1900)
 
-}  // namespace nodereport
+}  // namespace node-report
 
 #endif  // SRC_NODE_REPORT_H_