+++ CHANGES.md
@@ -1,7 +1,7 @@
 2016-12-12, Version 1.0.7
 =========================
 
- * Fix version reporting in NodeReport section (Richard Lau)
+ * Fix version reporting in nodereport section (Richard Lau)
 
  * Fix fprintf calls on Windows (Richard Lau)